requirements.txt
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.1
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.2
entrypoints==0.4
filelock==3.0.12
importlib-metadata==4.8.3
importlib-resources==5.1.4
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.3
jedi==0.17.2
Jinja2==3.0.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.2
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.6
notebook==6.4.10
numpy==1.19.5
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pipenv==2021.11.23
prometheus-client==0.16.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv
pytz==2022.7.1
pyzmq==25.0.0
qtconsole==5.2.2
QtPy==2.0.1
scipy==1.5.4
seaborn==0.11.2
selenium==3.141.0
Send2Trash==1.8.0
six==1.16.0
terminado==0.12.1
testpath==0.6.0
torch==1.10.2
tornado==6.1
traitlets==4.3.3
typing-extensions==4.1.1
urllib3==1.26.12
virtualenv==20.4.7
virtualenv-clone==0.5.7
wcwidth==0.2.6
webencodings==0.5.1
widgetsnbextension==3.6.2
zipp==3.6.0
openai==0.27.0
argparse
